+++ packages/deploy/src/lib.rs
@@ -1,4 +1,5 @@
 pub mod adodb;
+pub mod build;
 pub mod chains;
 pub mod contract_interface;
 pub mod contracts;
@@ -6,3 +7,4 @@ pub mod error;
 pub mod os;
 pub mod report;
 pub mod slack;
+pub mod validate;